Add a pass that allocates the local variables
This assigns each local variable a number that will be usable as an
offset from the stack pointer. The traversal of the CFG is structured
with a state monad that keeps a Map from variable names to numbers and
the next number to allocate. The one operation on the monad asks for the
number of a variable, and if it hasn't been seen yet, allocates a new
number for it, all in one operation.

This relies on ppx_monadic, so you must now
> opam install ppx_monadic
to compiler the compiler.

I'm not sure this is final yet. Live variable analysis works on the
string variable representation, and that's nice because we can get rid
of unused assignments before allocating offsets. However, we want
offsets before register allocation, and fancy allocators use the LVA
results. I could functorise LVA, but I'm not sure we want to teach that
on the mainline of the lectures. Probably, I'll move LVA after this
pass, and just live with possibly larger stack frames.